Point-load (per-column relative) loss term, and an autocast-safe decode
--column-weight adds the mean over sampled coordinates c of
||(M_hat - M)[:, c]||^2 / ||M[:, c]||^2: the point-load compliance error is
exactly this column's response, and the absolute bucket loss fits a column of
norm 15 (an ordinary face node) to ~15 % while fitting one of norm 2000 to
2-4 % (docs/data/softdiag_20260919/response_err).  Column norms come from the
same dense pass as the response bank and are cached with it; rows are
subsampled uniformly (count/R, unbiased); --column-tilt draws columns with
probability colnorm^-tilt.  Smoke on seat 253 (24 steps, response + column):
bank 3.4 s, the column term is the sixth reported bucket.

model.decode casts the head outputs to the accumulator dtype, which is what
torch.autocast needs (the LayerNorm-ed trunk state stays fp32 while the heads
emit bf16).  A step profile on seat 253 with two arms sharing the GPU: 136 ms
fp32 / 114 ms bf16 at 8192 pairs per bucket, 195 / 148 ms at 16384; backward
42-46 %, decode 21-26 %, loss 8-11 %, per-step rotation 7-9 %, sampling and
label reads 8 %.

+++ superelement/equi/model.py
@@ -280,8 +280,9 @@ class EquiModel(nn.Module):
         near = dist <= self.near_radius
         m = len(r_off)
         raw = torch.empty((m, 9), dtype=state.dtype, device=state.device)
-        raw[near] = .5 * (self.near_head(state[:m][near]) + self.near_head(state[m:][near]).reshape(-1, 3, 3).transpose(1, 2).reshape(-1, 9))
-        raw[~near] = .5 * (self.far_head(state[:m][~near]) + self.far_head(state[m:][~near]).reshape(-1, 3, 3).transpose(1, 2).reshape(-1, 9))
+        # .to(raw.dtype): under autocast the heads emit bf16 while the LayerNorm-ed state stays fp32
+        raw[near] = (.5 * (self.near_head(state[:m][near]) + self.near_head(state[m:][near]).reshape(-1, 3, 3).transpose(1, 2).reshape(-1, 9))).to(raw.dtype)
+        raw[~near] = (.5 * (self.far_head(state[:m][~near]) + self.far_head(state[m:][~near]).reshape(-1, 3, 3).transpose(1, 2).reshape(-1, 9))).to(raw.dtype)
         out = torch.zeros((len(rows), 3, 3), dtype=torch.float32, device=state.device)
         scale = torch.where(near, float(conditioning.same_patch_rms), float(conditioning.cross_patch_rms)).to(raw.dtype)
         out[off] = (raw * scale[:, None]).reshape(-1, 3, 3).float()
